tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-padding-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap: getComputedStyle().scrollPadding</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding">
<meta name="assert" content="scroll-padding computed value is per side, either the keyword auto or a computed <length-percentage> value.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<style>
  #target {
    font-size: 40px;
  }
</style>
<script>
test_computed_value("scroll-padding-top", "auto");
test_computed_value("scroll-padding-top", "0", "0px");
test_computed_value("scroll-padding-top", "10px");
test_computed_value("scroll-padding-top", "20%");
test_computed_value("scroll-padding-top", "calc(10px + 0.5em)", "30px");
test_computed_value("scroll-padding-top", "calc(10px - 0.5em)", "0px");
test_computed_value("scroll-padding-top", "calc(50% + 60px)");


test_computed_value("scroll-padding-right", "auto");
test_computed_value("scroll-padding-right", "0", "0px");
test_computed_value("scroll-padding-right", "10px");
test_computed_value("scroll-padding-right", "20%");
test_computed_value("scroll-padding-right", "calc(10px + 0.5em)", "30px");
test_computed_value("scroll-padding-right", "calc(10px - 0.5em)", "0px");
test_computed_value("scroll-padding-right", "calc(50% + 60px)");


test_computed_value("scroll-padding-bottom", "auto");
test_computed_value("scroll-padding-bottom", "0", "0px");
test_computed_value("scroll-padding-bottom", "10px");
test_computed_value("scroll-padding-bottom", "20%");
test_computed_value("scroll-padding-bottom", "calc(10px + 0.5em)", "30px");
test_computed_value("scroll-padding-bottom", "calc(10px - 0.5em)", "0px");
test_computed_value("scroll-padding-bottom", "calc(50% + 60px)");


test_computed_value("scroll-padding-left", "auto");
test_computed_value("scroll-padding-left", "0", "0px");
test_computed_value("scroll-padding-left", "10px");
test_computed_value("scroll-padding-left", "20%");
test_computed_value("scroll-padding-left", "calc(10px + 0.5em)", "30px");
test_computed_value("scroll-padding-left", "calc(10px - 0.5em)", "0px");
test_computed_value("scroll-padding-left", "calc(50% + 60px)");


test_computed_value("scroll-padding", "auto");
test_computed_value("scroll-padding", "10px");
test_computed_value("scroll-padding", "0", "0px");
test_computed_value("scroll-padding", "20%");
test_computed_value("scroll-padding", "calc(10px + 0.5em)", "30px");
test_computed_value("scroll-padding", "calc(10px - 0.5em)", "0px");

test_computed_value("scroll-padding", "1px 2px");
test_computed_value("scroll-padding", "1px 2px 3%");
test_computed_value("scroll-padding", "1px 2px 3% 4px");
test_computed_value("scroll-padding", "1px auto");
test_computed_value("scroll-padding", "0 0 0 0", "0px");
test_computed_value("scroll-padding", "auto auto auto auto", "auto");
</script>
</body>
</html>
